Changes to match server updates (#14)
* updated bitcoin to work with new server pattern

* updated to work with new AsserterTable location

* removed cargo patch + fmt

+++ src/responses/data/getblock.rs
@@ -13,11 +13,11 @@ use crate::responses::common::BitcoinTransaction;
 pub struct GetBlockResponse {
     pub hash: String,
     // confirmations: usize,
-    pub height: i64,
+    pub height: usize,
     pub version: usize,
     // versionHex: String,
     pub merkleroot: String,
-    pub time: i64,
+    pub time: usize,
     pub mediantime: i64,
     pub nonce: usize,
     pub bits: String,
